+++ DTCD-StorageSystem/src/Plugin.Meta.js
@@ -1,8 +1,7 @@
 export default {
-	type: 'core',
-	name: 'StorageSystem',
-	title: 'Система внутреннего хранения данных',
-	withDependencies: true,
-	version: '0.3.0',
-	priority: 3,
+  type: 'core',
+  name: 'StorageSystem',
+  title: 'Система хранения данных',
+  version: '0.3.0',
+  priority: 3,
 };