+++ DTCD-StorageSystem/tests/Plugin.test.js
@@ -1,27 +1,36 @@
-import { Plugin } from './../src/Plugin';
+import { StorageSystem } from '@/Plugin';
+import { SystemPlugin } from 'SDK';
+import { initApp } from 'utils/initApp';
 
-const pluginMeta = Plugin.getRegistrationMeta();
+initApp();
 
-describe('Check static method Plugin.getRegistrationMeta()', () => {
+test('StorageSystem extends SystemPlugin class', () => {
+  const storage = new StorageSystem('guid1');
+  expect(storage).toBeInstanceOf(SystemPlugin);
+});
+
+describe('Check StorageSystem static method getRegistrationMeta():', () => {
+
+  const pluginMeta = StorageSystem.getRegistrationMeta();
 
   test('Method exists', () => {
-    expect(Plugin.getRegistrationMeta).toBeDefined();
-    expect(typeof Plugin.getRegistrationMeta).toEqual('function');
+    expect(StorageSystem.getRegistrationMeta).toBeDefined();
+    expect(typeof StorageSystem.getRegistrationMeta).toEqual('function');
   });
 
-  test('Method returns a non-empty object ', () => {
-    expect(pluginMeta !== null).toEqual(true);
+  test('Method returns a non-empty object', () => {
+    expect(pluginMeta).toBeDefined();
+    expect(pluginMeta).not.toBeNull();
     expect(typeof pluginMeta).toEqual('object');
     expect(Object.keys(pluginMeta).length).toBeGreaterThan(0);
   });
 
-  const { type, name, title } = pluginMeta;
+  describe('Check returned meta object:', () => {
 
-  describe('Check returned meta object', () => {
+    const { type, name, title, version, priority } = pluginMeta;
 
     test('Property "type" is exists and is equal to "core" string', () => {
       expect(type).toBeDefined();
-      expect(typeof type).toEqual('string');
       expect(type).toEqual('core');
     });
 
@@ -33,6 +42,16 @@ describe('Check static method Plugin.getRegistrationMeta()', () => {
     test('Property "title" is exists and is a string', () => {
       expect(title).toBeDefined();
       expect(typeof title).toEqual('string');
+    });
+
+    test('Property "version" is exists and is a string ', () => {
+      expect(version).toBeDefined();
+      expect(typeof version).toEqual('string');
+    });
+
+    test('Property "priority" is exists and is a number', () => {
+      expect(priority).toBeDefined();
+      expect(typeof priority).toEqual('number');
     });
 
   })