+++ DTCD-StorageSystem/tests/modules/properties/_list.js
@@ -1,0 +1,7 @@
+import recordCount from './recordCount';
+import recordList from './recordList';
+
+export const propertyTestList = [
+  recordCount,
+  recordList,
+];