+++ DTCD-StorageSystem/tests/utils/initApp.js
@@ -1,0 +1,15 @@
+const systems = {
+  LogSystem: {
+    debug () {},
+    info () {},
+    warn () {},
+    error () {},
+    fatal () {},
+  },
+};
+
+export const initApp = () => {
+  global.Application = {
+    getSystem: (name) => systems[name],
+  };
+};