+++ DTCD-StorageSystem/tests/utils/systems.js
@@ -1,6 +1,0 @@
-export const systems = {
-  LogSystem: {
-    log () {},
-  },
-  EventSystem: {},
-};